ENH Updated README with important changes for Top 5 Project
This allows for sub tasks to be assigned to Top 5 issues without pulling them to the Top 5 group. We experience this when we had the Top 5 group as a root group and then another root group with issues we wanted to link to the Top 5 issue. By making all projects a subproject of Top 5 (or your preferred root project) we can freely link subtasks and get updates to the Top 5 issue.